move layout components to components/layout
