+++ template/web/app/buy-me-coffee/index.tsx
@@ -1,9 +1,9 @@
 'use client';
 
 import { useEffect, useState } from 'react';
-import Banner from '@/components/banner/banner';
-import Footer from '@/components/footer/Footer';
-import Header from '@/components/header/Header';
+import Banner from '@/components/layout/banner/banner';
+import Footer from '@/components/layout/footer/Footer';
+import Header from '@/components/layout/header/Header';
 import HorizontalLine from '@/components/horizontal-line/HorizontalLine';
 import BuyMeCoffeeContractDemo from './_components/ContractDemo';
 import Guide from './_components/Guide';
