+++ template/web/src/components/layout/banner/banner.test.tsx
