+++ template/web/src/components/layout/banner/banner.tsx
