+++ template/web/src/components/layout/header/AccountInfoPanel.tsx
